requirements.txt
jsonlines==1.2.0
numpy==1.23.5
pandas==1.1.5
Pillow==8.4.0
Pillow==10.4.0
protobuf==3.20.3
python-dotenv==1.0.1
Requests==2.32.3
torch==1.9.0
tqdm==4.61.2
transformers==4.9.1
